+++ src/hydrogen/duct_template.clj
@@ -6,9 +6,11 @@
   (:require [hydrogen.core.duct-template :as core]
             [hydrogen.session.cognito.duct-template :as cognito]
             [hydrogen.session.keycloak.duct-template :as keycloak]
-            [hydrogen.persistence.sql.duct-template :as sql]))
+            [hydrogen.persistence.sql.duct-template :as sql]
+            [hydrogen.scheduling.twarc.duct-template :as twarc]))
 
 (def core-profile core/profile)
 (def session.cognito-profile cognito/profile)
 (def session.keycloak-profile keycloak/profile)
 (def persistence.sql-profile sql/profile)
+(def scheduling.twarc-profile twarc/profile)