+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Added `+hydrogen/scheduling.twarc` profile, for job scheduling using `magnet/scheduling.twarc` Duct library.
+
 ## [0.3.9] - 2020-05-01
 ### Fixed
 - Added some new externs definitions to Cognito externs file (backported from one customer project)